% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeList.R
\name{makeTreeList}
\alias{makeTreeList}
\title{Make a Tree List}
\usage{
makeTreeList(df, root = df[1, 1])
}
\arguments{
\item{df}{a data.frame}

\item{root}{the entry in the data.frame that serves at the tree root node}
}
\value{
a list
}
\description{
Make a Tree List
}
\concept{geneSetPathwayAnalysis}
