% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeGeneSets.R
\name{mergeGeneSets}
\alias{mergeGeneSets}
\title{Merge Gene Sets}
\usage{
mergeGeneSets(
  geneSets,
  name = "merged",
  idConversion = FALSE,
  returnCounts = FALSE,
  forceCache = FALSE
)
}
\arguments{
\item{geneSets}{a list of vectors (e.g. gene sets)}

\item{name}{the name for the merged gene set}

\item{idConversion}{convert IDs from Uniprot to HGNC (Default: FALSE)}

\item{returnCounts}{changes return to a list where one entry is "counts" (Default: FALSE)}
}
\value{
a list of vectors with one entry (i.e. a gene set) with the given
  name
}
\description{
Merge Gene Sets
}
\examples{
tmp <- list(a=c("abc", "def", "deg", "ghi"), b=c("def", "deg", "ijk"))
mergeGeneSets(tmp, "y")

}
\concept{geneSetPathwayAnalysis}
