% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePatternsFromGeneSets.R
\name{removePatternsFromGeneSets}
\alias{removePatternsFromGeneSets}
\title{Remove entries matching a vector of patterns from a list of vectors}
\usage{
removePatternsFromGeneSets(geneSets, patterns, verbose = FALSE)
}
\arguments{
\item{geneSets}{a list of vectors (e.g. gene sets)}

\item{patterns}{a vector of patterns for grepl()}

\item{verbose}{a boolean whether to show debugging output}
}
\value{
the gene sets (list of vectors) without entries matching patterns
}
\description{
Remove entries matching a vector of patterns from a list of vectors
}
\examples{
tmp <- list(a=c("abc", "def", "deg", "ghi"), b=c("def", "deg", "ijk"))
removePatternsFromGeneSets(tmp, "def", TRUE)

}
\concept{geneSetPathwayAnalysis}
